Create README.md
This machine learning project aims to predict the presence of trauma in the records of 250,0000 individuals.  The dataset includes various health metrics and lifestyle indicators such as age, BMI, smoking habits, alcohol consumption, and the prevalence of certain health conditions. With trauma present in approximately 48.9% of the rows, the dataset is considered balanced in regard to project objectives.  We will initially compare two machine learning models—Random Forest and XGBoost—to determine which model performs better for this classification challenge. Based on the initial results, the best-performing model will be further optimized to achieve higher accuracy and reliability in predicting trauma. The dataset has been thoroughly cleaned, with no missing values or outliers, ensuring reliable model training and evaluation.